[devbox] Download nixpkgs using fetchTarball (#2639)
## Summary
Download nixpkgs using `fetchTarball`. We were previously using `fetchGit`, but it's too slow – users have been complaining that the first call to `devbox shell` takes to long, and part of it is due to `fetchGit`. Replacing with `fetchTarball` instead.

## How was it tested?
Re-built and ran on example projects.

## Is this change backwards-compatible?
Yes, I kept the same version of nixpkgs and everything.