[devbox] Add version command (#2652)
## Summary
Add `devbox version`. It prints the version that is set via ldflags (via
goreleaser). I tried to make BuildInfo work ... but couldn't. I think
for now ldflags are more reliable. That said, I restricted the ldflags
to the same subset that we'll get via BuildInfo) so it should later be
easy to switch between them.

## How was it tested?
Built locally and ran.

## Is this change backwards-compatible?
Yes.