boxcli,devbox: move default prompt into devbox.json
We currently set the devbox shell prompt in the shellrc template with:

	export PS1="(devbox) $PS1"

This breaks the prompt for users that have a theme or use a shell that
sets the prompt in some other way. To allow user's to opt out of the
prompt, move it into devbox.json where they can modify or delete it.
Running `devbox init` will now generate a config that looks like:

	{
	  "packages": [],
	  "shell": {
	    "init_hook": [
	      "# Here you can remove or customize the prompt for your project's devbox shell.",
	      "# By convention, individual users can set DEVBOX_NO_PROMPT=1 in their shell's",
	      "# rc file to opt out of prompt customization.",
	      "if [ -z \"$DEVBOX_NO_PROMPT\" ]; then",
	      "\tPS1=\"(devbox:tmp.wGAYMHRO) $PS1\"",
	      "fi"
	    ]
	  }
	}

This also means that users with an existing devbox config will no longer
get a prompt. If they want to get the default prompt back, they can
run:

	devbox init -fix prompt

This fix will automatically append the default devbox prompt to the
shell init hook unless there's already a command that contains `PS1=`.
A future change will print a hint when launching a shell to tell the
user about this change.

Note that the `-fix` flag accepts a comma-separated list of fixes in
case we need to add more as the config evolves. For simplicity, the
current implementation only allows "prompt", but can be redone later to
support more.