[Rust Planner] part 1: devbox shell (#97)
## Summary

This PR adds basic support for Rust, focussing on `devbox shell`. The
next PR will support `devbox build`.

**Rust nix package**
This PR uses the [Oxalica/rust
overlay](https://github.com/oxalica/rust-overlay). This has prebuilt
binaries which can be installed for the rust toolchain. This is an
improvement over mozilla's rust-overlay, but that involves compiling the
sources upon installing the nix packages, which is not optimal.

The [nixpkgs manual](https://nixos.org/manual/nixpkgs/stable/#rust)
suggests using a community overlay when one needs to pick different rust
versions (as we do). There's also
[enthusiasm](https://gutier.io/post/development-using-rust-with-nix/)
for this approach in the community.

An alternate approach is to use the standard rust toolchain approach.
This involves installing `rustup` and doing `rustup default <version>`
to change versions. The downside of this is that we lose the benefits of
the nix-store: The specific rust version's `rustc` and `cargo` binaries
are dynamically downloaded.

**Implementation**
As part of this, I needed to install support for overlays in the
generated `shell.nix`

## How was it tested?

adding cases to `testdata/` for rust `stable` (currently, version
`1.63.0`) and version `1.62.0`

**devbox shell**

```
> go install ./cmd/install

> cd testdata/rust/rust-stable
> devbox shell
> rustc --version
rustc 1.63.0 (4b91a6ea7 2022-08-08)



> cd testdata/rust/rust-1.62.0
> devbox shell
Installing nix packages. This may take a while...
Starting a devbox shell...
info: using existing install for '1.62.0-x86_64-apple-darwin'
info: default toolchain set to '1.62.0-x86_64-apple-darwin'

  1.62.0-x86_64-apple-darwin unchanged - rustc 1.62.0 (a8314ef7d 2022-06-27)

(devbox)
> rustc --version
rustc 1.62.0 (a8314ef7d 2022-06-27)
(devbox)
```