devbox: allow shell.init_hook to be a JSON array
Add a `ConfigShellCmds` type that supports marshalling/unmarshalling
shells commands from devbox.json as either a string or an array of
strings. This makes writing multi-line shell hooks a bit easier.

Example:

	{
	  "packages": [
	    "go"
	  ],
	  "shell": {
	    "init_hook": [
	      "echo 'line 1'",
	      "echo 'line 2'"
	    ]
	  }
	}

The original format is preserved so that if the hook is a string, it
will marshal back to a string. Converting from a string to an array
results in an array where the only element is the original string.