ci: add debug workflow
Add a debug workflow that makes it easier to figure out what's going
wrong in CI.

This workflow can only be started manually. It prints some diagnostic
info (like current rate limits) and then starts a tmate session. The
tmate session lets whoever started the run login into the runner with an
an interactive shell. A timeout ensures that the session is eventually
terminated.