Rename CICD workflows to be more descriptive (#328)
## Summary
Rename CICD workflows to be more descriptive. I was getting confused
with the meaning of each workflow based on the filename. Adopting a
naming convention of `<artifact>-<workflow>` so that all the worflows
related to the same artifact are next to each other (i.e. `cli`
worfklows are together, `doc` workflows are together and so on)

## How was it tested?
Can't test until I merge this PR