[php] Remove v1 planner, add composer definition to v2 (#359)
## Summary

Closes #356 

This removes PHP v1 planner (which was conflicting with v2 planner) and creates a definition for composer that ensures it uses the same PHP version that has the extensions added by the user. The easiest way to observe this is via example:


If user adds `php82`, `php82Packages.composer`, and `php82Extensions.imagick`

This generates the definition:

`php82 = pkgs.php82.withExtensions ({ enabled, all }: enabled ++ (with all; [ imagick ]));`

This means php has imagick but composer will use another version of php82 without imagick. This PR fixes that by adding the definition:

`php82Packages.composer = php82.packages.composer;`

this overwrites the composer package with one that uses the `php82.withExtensions` build.

## How was it tested?

* Tested all repro steps from https://github.com/jetpack-io/devbox/issues/356
* Added a few ext requirements to a `composer.json` (imagick, memcached) and then tried `composer install` and saw failure.
* Added extensions to devbox.json, shelled in and tried again and succeeded.
* For sanity, inspected generated shell.nix file to ensure definitions were correct.