openssh: update SSH config when DEVBOX_GATEWAY is set
Update ~/.config/devbox/ssh/config to include the value of
DEVBOX_GATEWAY when it's set. This makes debugging easier by clearing
out previous host keys and disabling strict host key checking. It also
creates an alias of gateway.devbox-debug so that:

	ssh gateway.devbox-debug

will always connect to whatever is in DEVBOX_GATEWAY.