[cloud shell] respect gitignore file in --config directory to limit large data syncs (#387)
## Summary

I discovered that the rust-stable project was not syncing, because it
stores compiled
artifacts in a child-directory: `.rustup`. This directory is in the
`.gitignore` file
but mutagen was syncing it.

In the long run, we should not sync any files that are gitignored
(agree?).
However, code for that is a bit more involved, so as a first pass I
implement
looking for a .gitignore file in the `configDir` and passing its rules
to mutagen
as `--ignore` flags. Thankfully, mutagen's ignore rules match gitignore
format.

## How was it tested?

was able to start `devbox cloud shell` in the `rust-stable` example
project