[cleanup] rename configDir to projectDir (#445)
## Summary

`configDir` is used to mean the directory-path corresponding to the
`--config`
flag in the `devbox` commands. Unfortunately, this can be conflated with
the
directory at `~/.config/devbox`, and so we can choose to rename it to
`projectDir`.

Future PRs will address:
1. changes in `cloud.go` and other `devbox cloud` code.
2. changes in plugin framework, where this is referred to `rootDir`.

## How was it tested?

tests pass
could open `devbox shell` in a golang example project.