[cleanup] plugin framework: rename rootDir to projectDir (#448)
## Summary

For the plugin framework, this renames rootDir to projectDir. This
change
makes it consistent with the rest of the devbox codebase.

did not change this one, because it may break users. Or are we early
enough that
we can change it?
```
docs/app/docs/devbox_examples/servers/apache.md:HTTPD_DEVBOX_CONFIG_DIR={PROJECT_DIR}
```

## How was it tested?

automated tests