[plugins] Join hash -r and source message and show in yellow (#452)
## Summary

This improves the message we show when a package that has env vars is
installed. It shows it in yellow at the very end always. It also joins
the `hash -r` command so that it's a single command.

This is a stopgap until we have a better way to use update env variables
while in shell. Once that is done, some of this change can be reverted.
(we might also be able to remove the need for `hash -r` command by
running it directly from devbox and/or only showing it when it actually
needs to be run)

I also considered creating a global env file that sources all the env
files but this would cause potential overwriting I didn't want.

## How was it tested?

<img width="595" alt="image"
src="https://user-images.githubusercontent.com/544948/211695932-63e628e4-3cb9-42a2-aab3-aed933930597.png">