Disallow script names with whitespace in them (#500)
## Summary

We shouldn't have allowed this in the first place, but if we allow it,
then it's impossible to support arbitrary commands, as well as passing
args to a script.

Improves error messages a bit by including the script name.

## How was it tested?
Added spaces to some scripts in `devbox.json` and called `devbox run`

### NOTE: this is a breaking change for any devbox.json with scripts
that have whitespace. Since the fix is trivial, I thought it's
reasonable to release it without backwards-compatibility.