plansdk: use nixpkgs mirror when in cloud VM  (#521)
This change is a bit hacky. I'd like to eventually clean this up to work
better with the various nix commands. It also reverts the previous env
var change since it's no longer needed.

When generating the various *.nix files, check to see if the cloud
mirror is reachable. If it is, download nixpkgs archives from there
instead of GitHub.

Also update the default commit hash to be the current nixpkgs unstable,
which is also in the cache. This guarantees that any new devbox projects
created after this release will hit the cache in the cloud.