[RFC] make Telemetry UserID be a uuid string (#574)
## Summary

**Motivation**:
From inspecting the segment logs, I am reminded that the userID is
logged as a 64 char.

Segment says that ideally this userID is something we track in our main
database. Longer term, I'm not sure we'd want our primary userIDs to be
TEXT kind in Postgres data type terms, compared to smaller uuid or int
variations.

**Approach**:
This PR amends the `UserIDFromGithubUsername` function to
__deterministically__
return a UUID string (or empty string if not authenticated or error
occurs).

It is important that the same UUID is returned for the same github
username.

**RFC: Breaking Change**
Tradeoff: in Segment we'll not be able to track the same users' activity
prior to this PR.

## How was it tested?

copied this function into a devbox-project for golang and executed it
multiple times
with different usernames.