devbox: unset TMPDIR in shell/run environment (#691)
Nix sets TMPDIR to its own path when building an environment and then
prints it with `print-dev-env`. Devbox "leaks" TMPDIR, meaning it takes
the TMPDIR from the parent shell and overwrites the Nix TMPDIR. However,
this doesn't work when TMPDIR isn't set in the parent shell, which is
the case in WSL and Docker.

This is a quick fix for just TMPDIR. We should double check the other
variables that `nix develop` unsets when there's more time to test.