[lagobot] clear .devbox dirs (#833)
## Summary

**Problem:**
If we have a `.devbox` directory in an example project, then running the
testscript
seems to fail. Something about copying that `.devbox` directory is
breaking some logic.
We should clear the `.devbox` directory before we execute the
testscript.

**Approach:**
In this PR, I clear the `.devbox` directory in the example project
__before__
copying the contents over into the testscript's workdir.

This has the downside that if we have a `devbox shell` active for that
project,
then its `.devbox` directory will get cleared out under it. This is
sub-optimal
but also not a common scenario.

I tried deleting the `.devbox` directory __after__ copying it into the
testscript's workdir,
but that failed. I think but couldn't confirm that it was due to some
permissions issue.

## How was it tested?

for `examples/development/csharp/hello-world`:
1. ran `devbox run run_test` to generate the `.devbox` dir
2. ran `DEVBOX_DEBUG=0 DEVBOX_EXAMPLE_TESTS=1 go test -v -run
TestExamples/development_csharp -count 1 ./testscripts`
- also, manually had to filter for `csharp` in the code because the
filtering happens _after_ we copy all the projects into the testscript
workdirs. I'll send another PR to fix this.