telemetry: improve stack traces, error names, contexts (#844)
Clean up and fix telemetry so that it can be actually useful. This
commit is rather large because telemetry happens to touch a lot of
different things.

Example of what it looks like in Sentry:
https://jetpack-io.sentry.io/issues/4047374127

Client
------

- Remove unnecessary telemetry configuration, such as `InitOpts`. Make
the `build` package the source of truth for version information and
link-time variables. Generate a global execution ID once at startup.
- Standardize the contexts to use names and values recognized by Sentry.
- Remove arbitrary CLI args and DEVBOX_ environment variables from the
context, as these were leaking information.

Errors
------

- In combination with the `redact` package, errors can now include data
that has been marked as safe for telemetry.
- Errors that haven't been marked as safe have a more helpful
placeholder containing the chain of error types up until the first
exported error.

Stack Traces
------------

- Use the correct file and function names. This allows Sentry to group
errors correctly.
- Don't record every wrapped error. This doesn't make sense in Go since
an error is a value containing the entire message. Otherwise you end up
with a bunch of error messages that repeat themselves.