nixstore: Nix expression for generating a nixpkgs index
This expression generates a JSON blob containing the metadata for every
derivation in nixpkgs, similar to that of hydra's `packages.json`. It
differs in that it also includes the store path for each package and
contains the system that generated the index. The system field lets us
know which packages can actually be evaluated for a given system, which
tends to differ from the systems that claim to be supported in the
metadata.

This isn't used anywhere yet, but it can be manually called as a script
to generate the JSON for prototyping or debugging.