nixstore: add support for remote stores
Add a `Remote` function to access a remote store as opposed to a local
one. For example:

	store, err := nixstore.Remote("https://cache.nixos.org")

creates a store that queries packages in the Nix binary cache. It
accesses packages using the same `S3BucketFS` implemented in `nixed`.
Resolving dependencies is done by looking at the `.narinfo` file for a
package.

Add test for remote store