[lockfile] Add lockfile and use search endpoint to resolve versions (#936)
## Summary

This introduces a simple lock file and uses the search service to
resolve versions. The lock file is created/updated whenever a package is
added/removed. If a package is manually added to devbox.json the
lockfile is updated on run/shell/install/shellenv.

Packages that don't have `@` are never added to lockfile.

This is behind a feature flag.

## How was it tested?

```bash
devbox add python@3.9
devbox run "python --version"
devbox rm python@3.9
```

* Inspected nix profile to ensure packages are added and removed as
expected.
* Inspected lock file to ensure contents are as expected.