[version update] use LAUNCHER_PATH in triggerUpdate function (#975)
## Summary

For trigger update to work, we need to invoke the "launcher" and not the
"devbox CLI binary".

To do so, we can use the `LAUNCHER_PATH` env-var. And fallback to
"devbox" if in PATH,
or finally as a fail-safe to the prior `os.Executable`.

## How was it tested?

:eyes:

carefully pair-programmed with @mikeland86. We'll test after building
the pre-release
in `0.4.9-dev`.