[update] Implement devbox update (#978)
## Summary

Implements `devbox update [pkg]...`

It basically just re-resolves version. If there's a new version that
satisfies constraints, it installs it and removes the previous one.

Other improvements:

* implemented `findPackageByName` which allows us to use `devbox update
go` instead of `devbox update go@1.20`. In follow up we can also use
this for `devbox rm`
* flake.nix optimization: If there are no dev packages (this is a legacy
planner concept that is almost never used) and we add at least one
nixpkgs flake input, we use that input for mkShell. This avoids having
to download an additional nixpkgs.


## How was it tested?

* Manually modified lockfile to point to older versions and ran `devbox
update` and `devbox update go`
* Inspected flake.nix 
* Inspected nix profile