[update] Silence update notice for devbox-global-shellenv and development-devbox (#982)
## Summary

For `devbox global shellenv` which may be used in the user's shellrc
file, its
not ideal to show the update notice everytime the user opens their shell
(likely
for a non-devbox purpose).

For devbox engineers, we need not see the notice everytime we execute a
devbox command
during development.

## How was it tested?

Added `fmt.Fprintf(os.Stderr, "version was checked")` in
`boxcli/root.go` just prior
to calling `vercheck.BuildVersion`. Then did `devbox global shellenv 1>
/dev/null`,
and saw that "version was checked" was not printed. As a sanity test,
verified
that it was printed for `devbox shellenv 1> /dev/null`.

Also ran `devbox run build` and saw that the notice was not printed
since I'd built
the `devbox` originally with the code of this PR.