[bugfix, cleanup] Finish moving plugin services to process compose (#1150)
## Summary
Removes `services` stanza from the plugin's json definition, in favor of
using `process-compose.yaml`.

* Fixes bug where, in doing `devbox services ls`, we would only return
the _first_ service defined in a plugin, and would call it the name of
the plugin (instead of the name of the service). This had some
downstream effects as well.
* Removes creation of wrappers of the form `<svc>-start|stop-service`.
They were using the json-based definitions of the services, but `devbox
service start <svc>` was using the process-compose definition, so they
could diverge. Some wrappers were not being generated either, due to the
bug above.
* Deletes obsolete code.
* Updates plugins that define services and bump their versions. Since
plugins are packaged in the CLI, and the process-compose yamls were
already being written before this change, I think this change is
backwards-compatible when it comes to plugins.

Reviewers, please advise on whether this change would affect service
port-forwarding.

## How was it tested?
Ran `devbox services ls|start|stop|up` with different plugins installed
and verified they worked as expected.

## Backwards-compatible?
NO, because we're removing the service wrappers (e.g.
`redis-start-service`). I don't think they're being used, but
@mikeland86 please confirm? If needed, I can replace them with `devbox
services start redis` or something like that.