[cleanup] introduce filegen package, and move functions that generate files into it (#1189)
## Summary

This PR introduces a `filegen` package. It will contain functions
related to files that devbox generates in `.devbox/gen`:
- to enable `devbox.PrintEnv`, via `filegen.GenerateForPrintEnv`.
- to enable `devbox.RunScript` via `filegen.WriteScriptsToFile`.

My motivation was that I wanted to pull logic out of impl and plansdk
into a single package (and internalize parts of it so it is easier to
reason about or follow). Currently, we have logic for generating code in
.devbox/gen that is spread in different .go files in both impl and
plansdk.

No logic change. This is mostly consolidating code that has spread over
`impl/devbox.go` into a smaller package.


## How was it tested?

`devbox shell` worked.

- [x] testscripts.