nix: don't pass --experimental-features to nix commands
When running nix commands, devbox passes two flags to enable
experimental features:

1. `nix --extra-experimental-features ...`
2. `nix --option experimental-features ...`

The second flag causes issues because it overwrites the user's nix.conf.
It doesn't seem to be doing anything, so remove it and just use
`--extra-experimental-features`. There's also some cleanup to switch
all callers of `nix.ExperimentalFlags` to use `nix.Command` instead.
This function sets the correct flags and forwards SIGINTs to Nix
instead of just killing the process when devbox gets interrupted.

Details
-------

A specific bug caused by this flag is
https://github.com/jetpack-io/devbox/pull/1263. The DeterminateSystems
installer configures Nix without a build users group and instead
enables the `auto-allocate-uids` setting in `/etc/nix/nix.conf` on
Linux. If this setting gets disabled by devbox, then the Nix daemon has
no way of building flakes and fails.

The reason why `--option experimental-features ...` was added is because
we were still seeing errors about ca-derivations not being enabled
even though the `--extra-experimental-features ca-derivations` flag was
set. Overwriting the config setting seemed to fix it, but that might've
been a red herring.

The ca-derivations feature is unique in that it must be set in the Nix
daemon, not just the client. This is because it changes the schema of
the Nix database, requiring a restart of the daemon. The only reliable
way to enable it is to add it to `/etc/nix/nix.conf`.

Given that this flag is currently breaking some users and doesn't seem
to be having the intended effect, I think we should remove it and
figure out another way to ensure ca-derivations is enabled. Switching
to using the DeterminateSystems installer and printing errors telling
the user how to enable it are probably the best bet.