[devcontainer] Changed to single user nix in devcontainer (#1325)
## Summary
Multi user nix installation with creating a devbox user in dockerfile
causes an issue in devcontainer with Nix due to permission issues with
subdirectories nix creates in `/tmp/`. Changing to a single user
installation and handling everything under root user in docker seems to
fix the issue.

## How was it tested?
- `devbox init && devbox add hello`
- `devbox generate devcontainer`
- open vscode: `code .`
- in vscode command palette (cmd + shift + p) run `Dev Containers:
Rebuild Without Cache and Reopen in Container`
- confirm that devcontainer gets created successfully and vscode
switches to remote devcontainer environment.