[per-OS Packages] fix devbox update for unversioned packages with platform/excluded-platform (#1370)
## Summary

This PR fixes a slightly edge-casey workflow:
1. User on macOS tries to `devbox shell` with `glibcLocales` in
devbox.json. Note: unversioned package.
2. Is asked to run: `devbox add glibcLocales --exclude-platforms
x86_64-darwin`
3. Running that, prints a notice to run `devbox update`.
4. Runs `devbox update`.

This should work, but was previously broken for two reasons:
1. In `update.go`, we were calling `devbox.Add` with `nil` for platforms
and excluded_platforms.
2. In `devbox.Add`, we check `versionedPkg.ValidateExists()` prior to
converting the raw package name to the versioned-name.
Both are now fixed.

## How was it tested?

Ran the workflow described above. It's hard to write a testscript
unit-test, since
some packages build may or may not build on whichever system is running
the unit-test. And mocking is too complicated here.