[rm nixpkgs] make HEAD request to BinaryCache to ensure binary is cached (#1318)
## Summary

This PR makes a HTTP HEAD request to
`https://cache.nixos.org/<hash>.narinfo` to verify that the binary is
indeed cached.

If not, we fallback to the legacy path that is dependent on downloading
nixpkgs.

**Implementation note**
In `IsInBinaryCache`, I wanted to insert the HEAD request. However, this
function can be called in a loop for all packages, which is bad for
perf.

Instead, in this design, a new function `FillNarInfoCache(packageList)`
makes the HEAD requests concurrently, storing results in an in-memory
cache.

We _require_ that callers invoke `FillNarInfoCache(packageList)` prior
to invoking `IsInBinaryCache`. Failing this, `IsInBinaryCache` will
return an error. Not thrilled with this design, since it may lead to
inadvertent errors, but this code is pretty core and our test coverage
should expose any codepaths that call `IsInBinaryCache` without
`FillNarInfoCache`.




## How was it tested?

I could do `devbox shell` in a project that has `hello` package. I
inserted some print statements to verify that the HEAD
response was 200 and that the caching for the boolean value of
isNarInfoInCache was working.